Support line comments in lexer
These are introduced by 0 or more "openers" (the lexer can support
multiple comment styles at once) and run to the end of the
line. Comments are skipped along with whitespace at the start of token
lexing cycle. NOTE: the comment openers are stored in a
user-allocated, NULL-terminated array pointed to by the lexer
struct. This pointer can be (and is by default) NULL to denote no
comments shall be used.