Replace `.is_finished` flag with `.status` enum
The lexer can be in any one of four states: *ready*, *lexing*,
*finished*, or *finished (abnormal)*. When in one of the finished
states, the lexer stays in that state until `lxl_lexer_reset()` is
called. Otherwise, the lexer flips between the *ready* state (at the
start/end of a token) and the *lexing* state (whilst lexing a token),
entering the *finished* state upon reaching the end of its input (and
after emitting the last non-sentinel token). The *finished (abnormal)*
state is entered when the lexer encounters some fatal error from which
it cannot proceed, even if it has not yet reached the end of its input.