Add `.after_token_hook()` to lexer
This is called at the end of `lxl_lexer__finish_token()` and can be
used to inject arbitrary logic into the token finalisation step just
before the lexer returns the token. One use case of this could be to
replace certain error tokens with a non-error token or create a custom
error (instead of having to handle that logic in the parser). The hook
function is given access to the full state of the lexer and can even
modify it (but that may be ill-advised).